[test] Car 테스트 분리
